change long to int SimulationState for simulationSteps.

+++ src/alifeSim/Simulation/SimulationState.java
@@ -66,7 +66,7 @@ public class SimulationState
 		return status;
 	}
 		
-	public long getSimulationSteps()
+	public int getSimulationSteps()
 	{
 		return simulationSteps;
 	}
